Make code clang-10 ready (#2092)
I locally installed clang-10 on one of my laptops. Mostly, clang now warns about implicit int to float conversions. While adding static_casts makes the code more verbose, I think that this might make people more aware of potentially lossy casts.

clang-10 support is not yet part of CI, so for now, this is a one time change with (hopefully) no visible effects.

+++ src/benchmarklib/tpcc/procedures/tpcc_new_order.cpp
@@ -169,7 +169,7 @@ bool TPCCNewOrder::_on_execute() {
     }
 
     // Calculate price of line item (OL_AMOUNT)
-    const auto ol_amount = order_line.ol_quantity * i_price;
+    const auto ol_amount = static_cast<float>(order_line.ol_quantity) * i_price;
 
     // Add to ORDER_LINE
     // TODO(anyone): This can be made faster if we interpret "For each O_OL_CNT item on the order" less strictly and
